+++ relations/Cargo.toml
@@ -24,10 +24,12 @@ ark-std = { version = "^0.3.0", default-features = false }
 blake2 = { version = "0.9.2", default-features = false }
 poseidon = { path = "../poseidon", default-features = false }
 
+criterion = "0.3"
+
 [dev-dependencies]
 tracing = { version = "0.1", features = [ "attributes" ] }
 tracing-subscriber = { version = "0.2" }
-criterion = "0.3"
+
 
 [[bench]]
 name = "bench"
